network: pod IPs should be stored in a cache file for binding plugin
Otherwise, the status of the VMI will contain only the GA IPs.

This fix was done because of a bug discovered in passt binding plugin
tests.
In passt the guest IP6 is not the same as the pod one (in case the guest
has not DHCPv6 client).
The fact that the pod IPv6 wasn't cached, caused the internal GA IPv6 to
be in the VMI status.

Signed-off-by: Alona Paz <alkaplan@redhat.com>