Remove console.LinuxExpecter
This was introduced in 383027016867d3140153c128b9fa0413c905f7f0 to wait
for EFI to finish booting before continuing.

The persistent backend storage tests are the only place that it was
used.  Could not get those tests to succeed at the commit where it was
introduced, nor any future tags.

Other "waiting for boot up" tests only rely on console.LoginTo* and not
this function.  Not sure that the LinuxExpecter is necessary when other
tests wait for a login prompt.

Signed-off-by: Brad Nabholz <bnabholz@gmail.com>