Abstract cmd client
As of now we directly search the filesystem
to find the socket and then initialize
a client for the command server in the launcher.

This is harder to test and change than it could be.

This commit encapsulates this into a simple yet effective
abstraction. The next step will be to use this abstraction
in all places that require the command client.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>