Avoid reporting error if the socket file got closed
When a shutdown of the guest OS got
triggered from the guest,
the socket file for the serial console
device could get removed even before virt-launcher-monitor
detects that the VM is down and made virt-tail aware.

Due to that, virt-tail should not exit with an error
when the socket file for the serial console gets removed.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>