Remove legacy command sockets
Remove handling of old/legacy command server sockets.
Previously we shared a host directory between the handler
and launchers. Through this directory we have shared a
socket for the command server.

This was obsoleted in v0.29, we now don't share sockets
through host directory. Rather we find an empty dir of
the launcher Pod where the socket is created.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>