refactor: rename `WithContainerImage`to `WithContainerDisk`
`WithContainerDisk` is a better name.
The method was extended to accept the container disk name parameter;
in this way there users can define multiple container disk with
different names.
Giving the possibility to chose the name, allow users to add
expectations on the vmi spec. For example:
```
Expect(disks[0].Alias.GetName()).To(Equal(<sustom_disk_name>))
```

Signed-off-by: fossedihelm <ffossemo@redhat.com>