Add and adjust UT, remove time consuming test
The test simulates a scenario which is not very important
We can assume CDI should be installed with kubevirt
if user want to use datavolumes.
Adjusted the webhook UT to accept VMs when cdi doesnt exist
and added UT to make sure we update conditions in case
we fail to create dv for vm.

Signed-off-by: Shelly Kagan <skagan@redhat.com>