[test] avoid testing with terminationGracePeriod=0 by default for VMIs
To speedup our test suite, in the past we configured
terminationGracePeriod=0 by default for all the VMIs
used in the functional tests.

This has a few potential implications that should
be carefully considered.

The major one is:
https://github.com/kubernetes/kubernetes/issues/120671 :
When a pod with spec.terminationGracePeriod: 0 is deleted without force,
(which is exactly what we are doing by default for our test VMIs!)
it's force-deleted from the API server, without kubelet killing
its containers and unmounting its volumes first.

This could be dangerous because:
- the deletion flow in the test is different from what probably
  99.9% of our users will rely on
- we are not really sure that the node resources got really freed
  up as expected before starting the next test introducing potential
  flakiness.

Setting by default spec.terminationGracePeriod: 1 on our test VMIs
should only barely slow down the test suite although ensuring
we are going to rely on the regular flow for pod deletion.

Extend one of the existing tests
("should be successfully started and stopped multiple times")
to ensure we are still also covering
the terminationGracePeriod=0 case.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>