[guest-log] Explicitly trap more exit signals on virt-tail
According to the CI results,
we have the feeling that the kubelet/CRI-O
is getting lazy (minutes) removing the virt-launcher
pod although guest-console-log container and virt-tail
process are clearly terminated at node level.
This is visible just with CRI-O >= 1.28 and
only when setting spec.terminationGracePeriodSeconds=0 at
the VMI.
Although this smells like a bug in the Kubelet/CRI-O
area, up to now we are able to hit this only for
the guest-console-log container so something there
is probably leading to some corner case.

Still not clear (at least to me) why,
it seems that we are not able to reproduce the issue
with this PR.

Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>