fix(ksm): set ksm-enabled label to true independently from node pressure
Currently, KubeVirt adds the `kubevirt.io/ksm-enabled` label to those nodes
that match the labelSelector defined in kvConfiguration. The value of that
label will be true only if that node is under pressure, false otherwise.

This is wrong since we need to provide users the information that a
particular node is configured to use ksm, so that it can
be used to schedule vms there.

Signed-off-by: fossedihelm <ffossemo@redhat.com>