tests/utils.go: Get rid of RunVMI
This function was giving the false impression of allowing the caller to
wait for the VMI to be running by allowing to supply a timeout.

In reality this function only created a VMI and was using a nonsensical
Eventually loop around the Create call.

Get rid of RunVMI by replacing all its usages with direct calls to the
kubevirt client.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>