Introduce a FeatureGate for SerialConsoleLog
We suspect that on recent (>=1.27.2)
CRI-O versions the guest-console-log container
could slow down the deletion of the
virt-launcher pods.
Flipping the default value for DisableSerialConsoleLog
on the Kubevirt' API (and enabling it back in the future)
is quite cumbersome so let's
introduce a feature gate (default off!) to
skip the guest-console-log container
from virt-launcher pod until we will be able to
get it working properly.

Signed-off-by: stirabos <stirabos@redhat.com>