Deactivate and activate passt interface after migration to renew the IP
In case the binding plugin is using "LinkRefresh" migration method,
the interface is set down and then up to renew the IP (cause the DHCP
client to re-request the network info).

Note: this behaviour will apply on any VM interface using a binding plugin
with "LinkRefresh" migration method - primary or secondary.

Signed-off-by: Andrei Kvapil <kvapss@gmail.com>
Signed-off-by: Alona Paz <alkaplan@redhat.com>