refactor: rename `NewRandomMigration` to `New`
Remove the `Random` part in the name function, since
there's nothing random. The only thing that is (almost)
random is the name, that is generated, but this pattern
is used in almost every resource in the testsuite (see vmi).
The `Migration` part can be removed since we are inside the
libmigration package.

Signed-off-by: fossedihelm <ffossemo@redhat.com>