[test] use a default value for VMI removal timeout
Today many tests are waiting up to a certain
timeout for a VMI to disappear but some of
them for 60 seconds, other 120, others again
for 180 and also for 240 seconds.
Let's use a single value across the whole
test suite and let's use
240 seconds to prevent us from
hitting: https://issues.redhat.com/browse/OCPBUGS-27949
In the future we will be able to reduce it
to 120 seconds with a one line change.

Signed-off-by: stirabos <stirabos@redhat.com>