e2e,guestlog: Reduce log rotate test runtime
The guestlog log rotate test takes on average over 7 and a half minutes
to run which makes it one of longest e2e tests.

The majority of this time is taken up with writing data to the log file
to force the log to rotate.

Reducing the amount of data written to the log and reducing the number
of expected rotated log files greatly reduces the test run time.

Signed-off-by: Brian Carey <bcarey@redhat.com>