tests, sriov: Drop IPv6 connectivity test
In a SR-IOV setup, the node and pod stack is not involved at all.
Only the guest stack is IP stack aware and that is not something
Kubevirt is interested to cover in the e2e tests.

An attempt to bump the guest OS has exposed this redundant dependency.
The new guest OS (Fedora) had issues to handle the IPv6 cloudinit
configuration when IPv4 was not set.

In order to continue checking the connectivity, IPv4 guest setups tests
have been left.

Signed-off-by: Edward Haas <edwardh@redhat.com>