virt-handler, virt-operator: Create ShadowNode object
The shadowNode object is created by the virt-handler pod that is
deployed on each node. The shadowNode's name is set to be the same as
the node's name.
the owner-reference of the shadow node is naturally the virt-launcher
pod that created it. For that the virt-handler's podUID and podName is
added to its manifest as env vars.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>