virt-handler/node-labeller: Refactor label logic
Currently, all label actions on the node-labeller occur on the node
object directly.
In order to consume this logic for the shadow node in future commits:
- separating removeLabellerLabels logic to a function that gets the
labels to remove, and another that then removes it from the node.
- Rename functions that specifically change the node in order to be more
specific.
- Move node patch creation to a function.
- Move semantic equality condition of node labels outside of patchNode
function in order to be used for shadownode patch in later commits.

The logic of the code is not changed.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>