virt-handler/node-labeller: Patch shadowNode object
This commit sets node-labeller to patch the labels to the shadowNode
instance.
The node is still being patched directly, in order to avoid labels
getting stale during the following upgrade scenario: during upgrade,
there is a time window where the new virt-handler is deployed, but
virt-controller is using the old version. This means that shadowNode is
populated by virt-handler's node-labeller, but it is not synced to the
node by the new shadowNode controller. In order to failing to label the
node during this period of time - virt handler will keep directly
patching the node alongside the shadowNode.

Unit tests are added to support shadow node cases.
node-labeller e2e test is updated accordingly.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>