tests, libnet, cloudinit: Replace assertions with `panic`
When building the network cloudinit spec, an error may occur.
Such an error is part of the spec build and depends on the cluster
health and possible bad input. None related to what tests are actually
verifying.

Therefore, use directly `panic` to differentiate that this is not a
test verification failure.

Signed-off-by: Edward Haas <edwardh@redhat.com>