virt-handler/heartbeat: Patch shadowNode object
This commit sets the heartbeat package labels and annotations to patch
the shadownode instance.
The node is still being patched directly, in order to avoid heartbeat
timeout during the following upgrade scenario: during upgrade, there is
a time window where the new virt-handler is deployed, but
virt-controller is using the old version. This means that shadowNode is
populated by virt-handler's heartbeat, but it is not synced to the
node by the new shadowNode controller. In order to avoid heartbeat
timeout during this period of time - virt-handler will keep directly
patching the node alongside the shadowNode.

Unit tests are added to support shadow node cases.
Heartbeat e2e test is updated accordingly.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>