virt-controller/node: Watch and patch shadowNode object
Currently the node controller works by watching the node for the
heartbeat, and set the  "kubevirt.io/schedulable" label on the node.

This commit moves the controller to watch the shadownode, as this is now
where the heartbeat is kept.
Is if needed it also patches the label to the shadowNode instead of
directly to the node, as the shadowNode controller now owns patching all
kubevirt related labels and annotations to the node.
Events are still sent directly to the node.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>