revert: Set domain firmware UUID based on namespace & name
The original change may break users when starting an old VM.
When starting with the new generated UUID, most likely the machine UUID
will change which in turn may place the guest in an unexpected state.

There is a need to distinguish old and new VMs in order to provide a
backward compatible fix.

Signed-off-by: Edward Haas <edwardh@redhat.com>