virt-handler/node: Patch shadowNode object
Currently the virt-handler sets the "kubevirt.io/schedulable" label on
the node if the main process exits.
The node is still being patched directly, in order to avoid setting the
lable during the following upgrade scenario: during upgrade, there is
a time window where the new virt-handler is deployed, but
virt-controller is using the old version. This means that shadowNode is
populated by virt-handler, but it is not synced to the
node by the new shadowNode controller. In order to avoid failing to
label the node during this period of time - virt-handler will keep
directly patching the node alongside the shadowNode.

This commit moves virt-handler to label the shadownode.

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>