test: Fix node taint test
Previously the test did not correctly check prefixes of pod names,
so it always passed. And after finishing, it left the cluster
in an inconsistent state, so other tests could fail.

This change adds an Eventually() to AfterEach() to wait
until all deployments that were running on the tainted node
are ready.

Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>