tests, network, services: Remove the VMI cleanup
VMI cleanup is already occurring during a test fixture teardown.
Most e2e tests are depending on this cleanup already, therefore it is
safe to expect no degradation will be observed.

Nonetheless, the current cleanup waits for the VMI objects to disappear.
While the generic VMI cleanup fixture is not.

Signed-off-by: Edward Haas <edwardh@redhat.com>