Fix quarantined clone and snapshot tests
AfterSuite failing becasue VM had "snapshot in progress" finalizer and could not be deleted.
Snapshot controller puts a finalizer on the VM while the snapshot is in progress.
VM finalizer should be removed when snapshot is complete but was not always happening when snapshot delete happened soon after success/failure

Three major issues addressed here:

1.  Make sure that Succeeded and Failed phases are terminal.  Flapping between phases was causing vmsnapshot finalizer to get removed too soon.
2.  Make sure to only remove vmsnapshot finalizer if vm finalizer was removed
3.  Namespace cleanup code should not remove finalizers from vmsnapshot objects.  Give the vmsnapshot controller time to do its job

Signed-off-by: Michael Henriksen <mhenriks@redhat.com>