Modify default garbage collection behavior in virtctl vmexport
When downloading a volume, the vmexport command always deletes the vmexport resource after a download.

After several conversations with the team, this behavior doesn't seem ideal since we could be deleting vmexport created in advance by other users.

This commit aims to change this default behavior so we only delete VMExport resources that've been created by the download command. It also adds a new flag to force-delete the vmexport anyway, mimicking the old behavior.

Signed-off-by: Alvaro Romero <alromero@redhat.com>