virt-handler: condition patching the node with SelfSubjectAccessReview
Although virt-handler now patches the shadownode, it also still patches
the node directly. This is done in order to avoid the following upgrade
scenario:
During upgrade, there is a time window where the new virt-handler is
deployed, but virt-controller is using the old version. This means that
the shadowNode is populated by virt-handler, but it is not synced to the
node by the new shadowNode controller. To avoid failing labelling the
node during this period of time - virt-handler will keep
directly patching the node alongside the shadowNode.

There is however the desire to reduce patching the node after the
upgrade is finished. In order to do so - this commit is adding a check
to see if virt-handler can actually patch the node before it does it.
This is done using the SelfSubjectAccessReview API [0].
By doing this, the node will no longer be patched as soon as the old
RBACs that previously permitted it are removed.

This commit
- adds a new package authorization to check if the node is patchable
- adds the appropriate RBAC to run the SelfSubjectAccessReview API on
virt-launcher.
- adds a check in cases where virt-handler attempts to patch the node,
and avoid the patch if it is forbidden. The cases are:
  - virt-handler/node-labeller.
  - virt-handler/heartbeat.
  - cmd/virt-handler.
- adjusts the unit tests accordingly.
- adds a new unit test to check the validity

[0]
https://kubernetes.io/docs/reference/kubernetes-api/authorization-resources/self-subject-access-review-v1/#SelfSubjectAccessReview

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>