virt-handler: Allow shadowNode CRD during upgrade
During the virt-operator upgradem CRD are deployed first in the upgrade
sequence [0]. However, new CRDs are blocked from being used until the
rollout is over [1].
During upgrade, the now newly introduced shadowNode CRD will be new, so
it would be blocked during the upgrade.
Since we need the shadownode CRD to be ready before the upgrade is over
in order to prevent heartbeat timeout, excluding the shadowNode CRD from
being blocked during upgrade.

[0]
https://github.com/kubevirt/kubevirt/blob/94c1ecd4dac289ffafea666d5b82e7553171ca1e/pkg/virt-operator/resource/apply/reconcile.go#L539
[1]
https://github.com/kubevirt/kubevirt/blob/94c1ecd4dac289ffafea666d5b82e7553171ca1e/pkg/virt-operator/resource/apply/reconcile.go#L341

Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>
Signed-off-by: Ram Lavi <ralavi@redhat.com>