migration: Check pod ready condition too
The live migration process wait for target pod readiness to close the
source pod, it checks the container statuses and the pod phase to
be running.

Sometimes the pod status is like the following:

```yaml
status:
    phase: Running                                                                conditions:
    - lastProbeTime: null
      lastTransitionTime: "2024-02-21T09:00:02Z"
      status: "True"
      type: Initialized
    - lastProbeTime: null
      lastTransitionTime: "2024-02-21T08:59:40Z"
      message: corresponding condition of pod readiness gate "kubevirt.io/virtual-machine-unpaused"
        does not exist.
      reason: ReadinessGatesNotReady
      status: "False"
      type: Ready
    - lastProbeTime: null
      lastTransitionTime: "2024-02-21T09:00:09Z"
      status: "True"
      type: ContainersReady
    - lastProbeTime: null
      lastTransitionTime: "2024-02-21T08:59:40Z"
      status: "True"
      type: PodScheduled
    - lastProbeTime: "2024-02-21T09:00:29Z"
      lastTransitionTime: "2024-02-21T09:00:29Z"
      message: the virtual machine is not paused
      reason: NotPaused
      status: "True"
      type: kubevirt.io/virtual-machine-unpaused
```

This means that Phase == Running but conditions.Ready == False.

At ovn-kubernetes CNI it means that the endpoint slices of a service
exposing the a service will be set to not ready and the ovs/ovn network
infra related to them will be deleted disconnecting the VM.

This change checks also the pod ready conditions to be sure that
endpoint slices readiness are aligned with successful live migraiton.

Signed-off-by: Enrique Llorente <ellorent@redhat.com>