network/tests: use NewGuestless instead of a complex vmi definition
The removed code started with NewRandomVMI and made an effort to erase
mostly everything there, rolling it down to something like what
NewGuestless is providing.

The removed code also added a Cirros boot disk, but it is not clear why.
The test does not expect the VM to boot at all.

I don't understand why we need this test at all, but I'm not addressing
this now.

Signed-off-by: Dan Kenigsberg <danken@redhat.com>