chore(fix): Reduces the downwardMetrics rate limit
The number of requests per second is set to 5, which causes an
excessive resources consumption by the virt-handler if many VMs are
running on the cluster and requesting the metrics.
It reduces the number of requests per second that can be addressed to
the downwardMetrics server to 1 to reduce the virt-handler resource
consumption.

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>