feat(client-gen): introduce CreateOptions for vmi.Create method
vmi.Create method does not allow to pass the CreateOptions.
In order to align the custom method signature with the generated one
introduce the possibility to pass the CreateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>