feat(client-gen): switch using ListOptions instead of pointer for vmi.List method
vmi.List method allow to pass the pointer to ListOptions.
In order to align the custom method signature with the generated one
switch using the ListOptions instead of the pointer one.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>