test(fix): virtiofs multiple PVCs tests
Virtiofs tests using multiple PVCs ask for a storage class supporting
access mode `ReadWriteMany`. For the purpose of these tests, it is not
required any specific access mode and, if this storage class is not
found, the PVC requested through this non-existent storage class gets
stuck in `Pending` state causing the test to fail.
It changes the requested storage class to any available one regardless
of the access mode. Moreover, this will ensure the test will work in
the majority of scenarios.

Signed-off-by: Javier Cano Cano <jcanocan@redhat.com>