virt-launcher, hooks: use libvirtxml.Domain
Hook sidecars that implement OnDefineDomain have the opportunity to do
changes that are not supported by virtwrap/api/schema.go hence, we
should be careful if calling the sidecars multiple times, in order to
have included in the XML, the previous changes made by the sidecar as
well.

By using libvirtxml.Domain we will keep stored the full changes for
the period we might call the sidecar twice, in the function
withNetworkIfacesResources.

The patch introduces in virtwrap/libvirtxml two convert functions, to
help convert between schema.go's DomainSpec and libvirtxml.Domain.

Fixes: 637be26ef4
Signed-off-by: Victor Toso <victortoso@redhat.com>