tests, net, slirp: Remove external connectivity tests
e2e tests that require external connectivity are fragile and heavily
dependent on the availability of the internet services and the endpoints
that are checked against. Therefore, these kind of tests should be
avoided.

Specifically, SLIRP is in the path of deprecation and has a very low
usage by users (practically we suspect no one uses it now-days).

Therefore, the SLIRP tests that cover external connectivity are dropped.

Signed-off-by: Edward Haas <edwardh@redhat.com>