feat(client-gen): introduce CreateOptions for vm.Create method
vm.Create method does not allow to pass the CreateOptions.
In order to align the custom method signature with the generated one
introduce the possibility to pass the CreateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

(cherry picked from commit 994fb9ef266a5db1317f5ee2ac359b040dffcbbe)

Signed-off-by: fossedihelm <ffossemo@redhat.com>