feat(client-gen): introduce UpdateOptions for vm.UpdateStatus method
vm.UpdateStatus method does not allow to pass the UpdateOptions.
In order to align the custom method signature with the generated one
introduce the possibility to pass the UpdateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit 3eba48518946b713764e6f09af4216f63f2229ef)