e2e: Double number of e2e parallel nodes from 3 to 6
Increasing the number of parallel nodes will reduce the runtime for e2e
test lanes that include parallel tests.

An increase in kubevirtci node memory is also required in order to benefit from
this increased parallelization. The new test clusters still fit within
the memory request limit of the test pods which is currently at 29GB.

Signed-off-by: Brian Carey <bcarey@redhat.com>