feat: use common params for vmim.List
vmim.List method does not allow to pass the context and uses the pointer to ListOptions.
In order to align the custom method signature with the generated one
switch using the ListOptions instead of the pointer one and introduce the context.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>