feat: use common params for vmim.Update
vmim.Update method does not allow to pass the context and UpdateOptions.
In order to align the custom method signature with the generated one introduce
the context and the UpdateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit 1696db440964a798f25647da0b7f36cd0b96cf37)