Eviction admitter: Deny when triggering VMI evacuation
Currently, when the criteria for VMI evacuation are met [1]:
1. The eviction admitter marks the VMI for evacuation.
2. The eviction admitter approves the eviction request.
3. kube-api denies the eviction request because the virt-launcher is
protected by a PDB [2].

When a descheduler[3] tries to evict a virt-launcher pod
owned by a VMI that could be evacuated,
it receives a denial for the eviction request and cannot understand
that a VMI evacuation was triggered in the background.

In order to support descheduler integration [4], deny the eviction
request with a spacial message when the VMI is marked for evacuation.
This is done in order to:
1. Signal the descheduler that an evacuation was triggered although
the eviction was denied.
2. Use the existing K8s eviction API.

In this scenario, kube-api will not check whether the virt-launcher is
protected by a PDB.

Note: In case the Eviction admitter is down - kube-api will treat the
eviction request as if it was approved by the admitter, and will continue to
check whether a PDB protects the virt-launcher.
In this case the VMI will not be marked for evacuation and the
eviction initiator receive a regular denial response because of the PDB.

[1] https://github.com/kubevirt/kubevirt/pull/11286
[2] https://kubernetes.io/docs/concepts/scheduling-eviction/api-eviction/#how-api-initiated-eviction-works
[3] https://github.com/kubernetes-sigs/descheduler
[4] https://github.com/kubevirt/community/pull/258

Signed-off-by: Orel Misan <omisan@redhat.com>