feat: use common params for vmim.UpdateStatus
vmim.UpdateStatus method does not allow to pass the context and UpdateOptions.
In order to align the custom method signature with the generated one introduce
the context and the UpdateOptions.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Also, removed duplicate metav1 import with different names.

Signed-off-by: fossedihelm <ffossemo@redhat.com>