tests, k8s: Improve log collection logging
68ebda2ae7d20 started to always collect logs when focus containers are
used by tests. While useful the step logging suggested that a failure
had occurred even if the test passes and was confusing to the reader.

This change provides a useful reason for the log collection and also
renames the associated KubernetesReporter attribute to make it clear
what it actually relates to.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>