controller: Move VMI CurrentCPUTopology defaulting to webhook
This change moves defaulting of CurrentCPUTopology in the VMI from the
VM controller into the VMI mutation webhook. This defaulting would
happen prior to instance types and preferences being applied to the VMI
and would thus capture in the incorrect state of the VMI when instance
types and preferences are used. The following changes in this series
require this in order to introduce support instance types and
preferences with LiveUpdate.

This also fixes bug #11251 where CurrentCPUTopology is not updated to
reflect a new instance type being used in clusters where LiveUpdate is
not currently enabled.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
(cherry picked from commit 01a37252ab4dc0539e1496dd751bcd711623794b)
Signed-off-by: Luboslav Pivarc <lpivarc@redhat.com>