controller: Move VMI hotplug setup to webhook
As with the previous change this change moves setup logic for hotplug
into the VMI mutation webhook in order to allow for instance type and
preference support of LiveUpdate.

By moving this logic into the mutation webhook we can now provide
MaxSockets and MaxGuest as centralised instance type attributes that are
applied to the VMI during creation.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>