api: Allow VMs to reference instance types with LiveUpdate enabled
As set out by bug #11960 this would previously cause any attempt to
create or update a VM referencing an instance type to be rejected with
the LiveUpdate feature enabled within the cluster.

While full LiveUpdate support for VMs referencing instance types is
still pending this block goes too far and should be removed to avoid
breaking existing VMs.

A follow up change will ensure that the RestartRequired condition is
added to VMs if their referenced instance type or preference changes
with LiveUpdate enabled.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>