migration controller: add a secondary queue to keep the main one managable
When starting a large number of migrations, most of them will stay pending for a long time.
Since pending migrations keep consuming controller execution cycles to just get re-enqueued,
there can be not enough cycles left for active migrations to move forward in time.

This commit introduces a second queue dedicated to the pending migrations.
This second queue will only be consumed when a migration finishes.

Signed-off-by: Jed Lejosne <jed@redhat.com>