Improve time interval in Eventually in tests/
The gomega's Eventually and Consistently function API is ambiguous. The
time interval parameters (timeout and polling) are both optional, and
can get multiple values, such as integers or floating points as number
of seconds, or time.Duration. This is very confusing.

This PR removes all these ambiguous parameters, replacing them with the
explicit versions: the `WithTimeout()` and `WithPolling()` methods, that
also only receives time.Duration values.

That make the usage in Eventually and Consistently more readable and
explicit.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>