Hash HostName value on DownwardMetrics
A concern was raised about the possibility of exposing the host name to
an untrusted VM via downward metrics. Let's return a hashed HostName
value to avoid exposing the real host name to the VM.

The HostName value is used by SAP as an uuid of the node (to detect
changes), and is not used to do dns queries.

Signed-off-by: German Maglione <gmaglione@redhat.com>