Eviction admitter, tests: Test VMI evacuation triggering
The main job of the pod eviction admitter is triggering
VMI evacuation.

This is done by patching the VMI's `Status.EvacuationNodeName`
to the node name it is currently scheduled on.

The evacuation request will be approved by the admitter
but will eventually fail because of a PodDistributionBudget
protecting the virt-launcher pod.

Create a new test to cover all the possible scenarios by setting:
- Cluster-wide eviction strategy
- VMI eviction strategy
- VMI migratable status

Remove old tests.

For additional details please see [1].

[1] https://github.com/kubevirt/kubevirt/pull/11286

Signed-off-by: Orel Misan <omisan@redhat.com>