Eviction admitter, tests: Deny and not trigger
Currently, the scenarios in which either the cluster wide, or the VMI
eviction strategy are set to `LiveMigrate` but the VMI is not migratable
are covered.

In these cases, the admitter denies the eviction with an error message.

Create a new test to cover all the possible scenarios by setting:
- Cluster-wide eviction strategy
- VMI eviction strategy

For additional details please see [1].

[1] https://github.com/kubevirt/kubevirt/pull/11286

Signed-off-by: Orel Misan <omisan@redhat.com>