feat: use common params for kv.Get
kv.Get method does not allow to pass the context and uses the pointer to GetOptions.
In order to align the custom method signature with the generated one
switch using the GetOptions instead of the pointer one and introduce the context.
This effort, once all the common methods' signatures are aligned with
the generated ones, will converge in the deletion of the
custom methods' implementations, embedding the generated interface.

Signed-off-by: fossedihelm <ffossemo@redhat.com>
(cherry picked from commit 190d2cf530e6b2bf592d7c8df7d31ffe3c3ed3f1)