virt-launcher: Stop DHCP server for hot-unplugged interfaces
Currently, the DHCP server for a hot-plugged interface is not
stopped after it is unplugged, the dhcpStartedFile is also not removed
so when an interface with the same name is removed and added back,
the corresponding DHCP server is not started and the interface
won't get configured.

This patch fixes this by watching NIC events and stop the DHCP server
after receiving a deletion event.

Signed-off-by: Jaff Cheng <jaff.cheng.sh@gmail.com>