network, network-info: generate network-info annotation
The VMI controller can use this method when generating the pod
annotations. The method doesn't retrun an error in case of marshaling
failure, so the VMI controller using it will just get an empty annotation.
If evetually the annotation won't be created, a follow-up step will fail,
there is no need to corrupt the vmi controller.


Signed-off-by: Alona Paz <alkaplan@redhat.com>